Revive AnyTopicsFilter for Baseline
